Move KodiakIsland price handler into berachain subgraph
