+++ subgraphs/berachain/src/price/PriceLookup.ts
@@ -1,7 +1,6 @@
 import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
 
 import { PriceHandler, PriceLookupResult } from "../../../shared/src/price/PriceHandler";
-import { PriceHandlerKodiakIsland } from "../../../shared/src/price/PriceHandlerKodiakIsland";
 import { PriceHandlerUniswapV3Quoter } from "../../../shared/src/price/PriceHandlerUniswapV3Quoter";
 import { getUSDRate } from "../../../shared/src/price/PriceRouter";
 import {
@@ -15,6 +14,7 @@ import {
   LP_UNISWAP_V3_WBERA_HONEY} from "../contracts/Constants";
 import { getContractName } from "../contracts/Contracts";
 import { getBaseTokenRate, isBaseToken } from "./PriceBase";
+import { PriceHandlerKodiakIsland } from "./PriceHandlerKodiakIsland";
 import { PriceHandlerKodiakIslandBeradrome } from "./PriceHandlerKodiakIslandBeradrome";
 
 const KODIAK_QUOTER = "0x644C8D6E501f7C994B74F5ceA96abe65d0BA662B".toLowerCase();
