+++ subgraphs/berachain/subgraph.yaml
@@ -25,9 +25,9 @@ dataSources:
         - name: UniswapV3Quoter
           file: ../shared/abis/UniswapV3Quoter.json
         - name: KodiakIsland
-          file: ../shared/abis/KodiakIsland.json
+          file: ./abis/KodiakIsland.json
         - name: BeradromeKodiakIslandRewardVault
-          file: abis/BeradromeKodiakIslandRewardVault.json
+          file: ./abis/BeradromeKodiakIslandRewardVault.json
         - name: UniswapV2Pair
           file: ../shared/abis/UniswapV2Pair.json
         - name: UniswapV3Pair
